--- src/frequenz/channels/util/__init__.py
@@ -1,36 +1,0 @@
-# License: MIT
-# Copyright © 2022 Frequenz Energy-as-a-Service GmbH
-
-"""Channel utilities.
-
-A module with several utilities to work with channels:
-
-* [Event][frequenz.channels.util.Event]:
-  A [receiver][frequenz.channels.Receiver] that can be made ready through an event.
-
-* [FileWatcher][frequenz.channels.util.FileWatcher]:
-  A [receiver][frequenz.channels.Receiver] that watches for file events.
-
-* [Timer][frequenz.channels.util.Timer]:
-  A [receiver][frequenz.channels.Receiver] that ticks at certain intervals.
-"""
-
-from ._event import Event
-from ._file_watcher import FileWatcher
-from ._timer import (
-    MissedTickPolicy,
-    SkipMissedAndDrift,
-    SkipMissedAndResync,
-    Timer,
-    TriggerAllMissed,
-)
-
-__all__ = [
-    "Event",
-    "FileWatcher",
-    "MissedTickPolicy",
-    "SkipMissedAndDrift",
-    "SkipMissedAndResync",
-    "Timer",
-    "TriggerAllMissed",
-]
